Rename Gaomon parser to UCLogic
