+++ OpenTabletDriver/Vendors/UCLogic/UCLogicReportParser.cs
@@ -1,15 +1,15 @@
 using OpenTabletDriver.Tablet;
 using OpenTabletDriver.Plugin.Tablet;
 
-namespace OpenTabletDriver.Vendors.Gaomon
+namespace OpenTabletDriver.Vendors.UCLogic
 {
-    public class GaomonReportParser : IReportParser<IDeviceReport>
+    public class UCLogicReportParser : IReportParser<IDeviceReport>
     {
         public IDeviceReport Parse(byte[] data)
         {
             var isAuxReport = ((data[1] & (1 << 5)) != 0) & ((data[1] & (1 << 6)) != 0);
             if (isAuxReport)
-                return new GaomonAuxReport(data);
+                return new UCLogicAuxReport(data);
             else
                 return new TabletReport(data);
         }
